silently ignore when disabling type checks for a file format that isn't supported
