+++ packages/instrumenter/src/disable-type-checks.ts
@@ -10,14 +10,27 @@ const commentDirectiveRegEx = /^(\s*)@(ts-[a-z-]+).*$/;
 const tsDirectiveLikeRegEx = /@(ts-[a-z-]+)/;
 const startingCommentRegex = /(^\s*\/\*.*?\*\/)/gs;
 
+/**
+ * Disables TypeScript type checking for a single file by inserting `// @ts-nocheck` commands.
+ * It also does this for *.js files, as they can be type checked by typescript as well.
+ * Other file types are silently ignored
+ *
+ * @see https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#-ts-nocheck-in-typescript-files
+ */
 export async function disableTypeChecks(file: File, options: ParserOptions): Promise<File> {
-  if (isJSFileWithoutTSDirectives(file)) {
+  const format = getFormat(file.name);
+  if (!format) {
+    // Readme files and stuff don't need disabling.
+    return file;
+  }
+  if (isJSFileWithoutTSDirectives(file, format)) {
     // Performance optimization. Only parse the file when it has a change of containing a `// @ts-` directive
     return {
       ...file,
       content: prefixWithNoCheck(file.content),
     };
   }
+
   const parse = createParser(options);
   const ast = await parse(file.content, file.name);
   switch (ast.format) {
@@ -30,8 +43,7 @@ export async function disableTypeChecks(file: File, options: ParserOptions): Pro
   }
 }
 
-function isJSFileWithoutTSDirectives(file: File) {
-  const format = getFormat(file.name);
+function isJSFileWithoutTSDirectives(file: File, format: AstFormat) {
   return (format === AstFormat.TS || format === AstFormat.JS) && !tsDirectiveLikeRegEx.test(file.content);
 }
 
@@ -44,7 +56,7 @@ function prefixWithNoCheck(code: string): string {
     // first line has a shebang (#!/usr/bin/env node)
     const newLineIndex = code.indexOf('\n');
     if (newLineIndex > 0) {
-      return `${code.substr(0, newLineIndex)}\n// @ts-nocheck\n${code.substr(newLineIndex + 1)}`;
+      return `${code.substring(0, newLineIndex)}\n// @ts-nocheck\n${code.substring(newLineIndex + 1)}`;
     } else {
       return code;
     }
@@ -52,7 +64,7 @@ function prefixWithNoCheck(code: string): string {
     // We should leave comments, like `/** @jest-env jsdom */ at the top of the file, see #2569
     startingCommentRegex.lastIndex = 0;
     const commentMatch = startingCommentRegex.exec(code);
-    return `${commentMatch?.[1].concat('\n') ?? ''}// @ts-nocheck\n${code.substr(commentMatch?.[1].length ?? 0)}`;
+    return `${commentMatch?.[1].concat('\n') ?? ''}// @ts-nocheck\n${code.substring(commentMatch?.[1].length ?? 0)}`;
   }
 }
